# openapi.yaml
...
openapi: "3.0.3"
info:
  title: sc-agent
  description: API to interact with sc-agent
  version: 1.0.0
  contact:
    name: soeren schneider
    url: https://github.com/soerenschneider/sc-agent
    email: "56670304+soerenschneider@users.noreply.github.com" # this is just for linting reasons
  license:
    name: "NONE" # this is just for linting reasons
    url: "none" # this is just for linting reasons
servers:
  - url: http://localhost:9999
    description: Local server
...
paths:
...
  /v1/certs/x509/issue-requests:
    post:
      operationId: certsX509PostIssueRequests
      summary: Issues a X509 certificate
      description: Issues a X509 certificate
      tags:
        - certs
        - x509
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
          example: "my-user-cert"
          description: The id of the SSH host certificate you want to sign
        - name: force-renewal
          in: query
          required: false
          schema:
            type: boolean
          description: Do not check the lifetime of the existing certificate against the threshold and force a new signature of the host key
      responses:
        '200':
          description: Certificate already exists and is still valid, no new signature was needed
        '201':
          description: New signature created, either because certificate did not exist, was expired, was below the lifetime threshold or because force_renewal was set to true.
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '501':
          $ref: '#/components/responses/NotImplemented'
  /v1/certs/x509:
    get:
      operationId: certsX509GetCertificatesList
      summary: Return configuration of all managed x509 certificates
      description: >
        Returns a list of all configured managed x509 certificates.
      tags:
        - certs
        - x509
      responses:
        '200':
          description: Configuration of all managed x509 certificates
          content:
            application/json:
              schema:
                type: array
                x-go-type-skip-optional-pointer: true
                items:
                  $ref: '#/components/schemas/X509ManagedCertificateList'
                description: List of managed certificate configs
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '501':
          $ref: '#/components/responses/NotImplemented'
  /v1/certs/x509/{id}:
    get:
      operationId: certsX509GetCertificate
      summary: Returns info of configured managed x509 certificate
      description: Accepts the id of the configured managed x509 certificate as path parameter and returns its config.
      tags:
        - certs
        - x509
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "my-id"
          description: The id of the managed x509 certificate you want to retrieve configuration for
      responses:
        '200':
          description: Info about the managed x509 cert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/X509ManagedCertificate"
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '501':
          $ref: '#/components/responses/NotImplemented'
...
  /v1/k0s/actions:
    post:
      operationId: k0sPostAction
      summary: Start k0s
      description: Start the k0s service.
      tags:
        - k0s
      parameters:
        - name: action
          in: query
          required: true
          schema:
            type: string
            enum:
              - stop
              - start
          example: "stop"
          description: Action to perform for the k0s service
      responses:
        '200':
          description: Command to start k0s was invoked successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '501':
          $ref: '#/components/responses/NotImplemented'
  /v1/packages/installed:
    get:
      operationId: packagesInstalledGet
      summary: Lists installed packages
      description: List all installed packages that are installed via the system's package manager.
      tags:
        - packages
      responses:
        '200':
          description: A list of packages installed via DNF/RPM on the system
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackagesInstalled'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '501':
          $ref: '#/components/responses/NotImplemented'
  /v1/packages/updates:
    get:
      operationId: packagesUpdatesGet
      summary: Checks for updates
      description: Checks if any packages are updateable via DNF
      tags:
        - packages
      responses:
        '200':
          description: Returns whether updates are available for packages installed via DNF
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackageUpdates'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '501':
          $ref: '#/components/responses/NotImplemented'
  /v1/packages/upgrade-requests:
    post:
      operationId: packagesUpgradeRequestsPost
      summary: Upgrades all packages
      description: Upgrades all packages with the default package manager. It is not recommend to use this endpoint on a regular basis.
      tags:
        - packages
      responses:
        '200':
          description: Upgrade was successful
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '501':
          $ref: '#/components/responses/NotImplemented'
  /v1/libvirt/domains/{domain}:
    post:
      operationId: libvirtPostDomainAction
      summary: Set status of libvirt domain
      description: >
        Interact with the status of a libvirt domain running on the machine the API runs on. Both domain name and
        action are set as path variable.
      tags:
        - libvirt
      parameters:
        - name: domain
          in: path
          required: true
          schema:
            type: string
          description: The name of the domain you want to shutdown
          example: example-domain
        - name: action
          in: query
          required: true
          schema:
            type: string
            enum:
              - reboot
              - shutdown
              - start
          example: reboot
          description: Action to perform for the libvirt domain
      responses:
        '200':
          description: Domain shutdown successful
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '501':
          $ref: '#/components/responses/NotImplemented'
...
  /v1/replication/http/items:
    get:
      operationId: replicationGetHttpItemsList
      summary: "Returns current configuration"
      description: Returns the current configuration of the secret replication component.
      tags:
        - secrets
      responses:
        '200':
          description: The current configuration of the secret syncer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReplicationHttpItemsList"
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '501':
          $ref: '#/components/responses/NotImplemented'
  /v1/replication/http/items/{id}:
    get:
      operationId: replicationGetHttpItem
      summary: "Returns the replication status of a single item"
      description: >
        Returns the replication status of a single secret replication item. The id of the item is passed via path
        variable.
      tags:
        - secrets
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "prod/db"
          description: The id which translates to the secret path of the to-be-synced item in Vault.
      responses:
        '200':
          description: Object containing the status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReplicationHttpItem"
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '501':
          $ref: '#/components/responses/NotImplemented'
...
components:
  schemas:
    Problem:
      description: A generic HTTP problem
      xml:
        name: problem
        namespace: urn:ietf:rfc:7807
      properties:
        type:
          type: "string"
          description: "A URI reference that identifies the problem type"
          example: "https://example.com/probs/out-of-credit"
        title:
          type: "string"
          description: "A short, human-readable summary of the problem type"
          example: "You do not have enough credit."
        detail:
          type: "string"
          description: "A human-readable explanation specific to this occurrence of the problem"
          example: "Your current balance is 30, but that costs 50."
        instance:
          type: "string"
          description: "A URI reference that identifies the specific occurrence of the problem"
          example: "/account/12345/msgs/abc"
        status:
          type: "integer"
          description: "The HTTP status code"
          example: 400
      example:
        type: "https://example.com/probs/out-of-credit"
        title: "You do not have enough credit."
        detail: "You do not have enough credit."
        instance: "/account/12345/msgs/abc"
        status: 400
...
    ReplicationHttpItem:
      type: object
      title: HttpReplicationItem
      description: "Configuration and status of a single HTTP replication item"
      properties:
        id:
          type: string
          example: "ca-crt"
          x-go-type-skip-optional-pointer: true
          description: id of the item
        source:
          type: string
          example: "prod/db"
          x-go-type-skip-optional-pointer: true
          description: path of the secret to read and sync to the local filesystem
        dest_uri:
          type: string
          example: "/tmp/db.json"
          x-go-type-skip-optional-pointer: true
          description: destination path where the read secret should be writen to
        expected_checksum:
          type: string
          example: 9f86d081884c7d659a2feaa0c55ad015a3bf4f1b2b0b822cd15d6c15b0f00a08
          description: only save the file if the checksum matches
        status:
          type: string
          example: "synced"
          x-go-type-skip-optional-pointer: true
          description: the status of the synced secret
          enum:
            - unknown
            - synced
            - failed
            - invalid_checksum
        post_hooks:
          type: array
          x-go-type-skip-optional-pointer: true
          items:
            $ref: '#/components/schemas/PostHooks'
    ReplicationHttpItemsList:
      type: object
      title: SecretReplicationItems
      description: "All managed HTTP replication items"
      properties:
        data:
          type: array
          x-go-type-skip-optional-pointer: true
          items:
            $ref: '#/components/schemas/ReplicationHttpItem'
          description: The sync items
      example:
        data:
          - secret_path: "prod/db"
            formatter: "json"
            dest_uri: "/tmp/db.json"
...
    PackageUpdates:
      title: PackageUpdates
      type: object
      description: "Returns whether system updates are available and which packages can be updated"
      properties:
        updates_available:
          type: boolean
          x-go-type-skip-optional-pointer: true
          description: Indicates if there are updates available
          example: true
        updatable_packages:
          type: array
          x-go-type-skip-optional-pointer: true
          items:
            $ref: '#/components/schemas/PackageInfo'
          description: List of packages that can be updated
      example:
        updates_available: true
        updatable_packages:
          - name: "nginx"
            version: "1.18.0"
            repo: "main"
    PackageInfo:
      type: object
      title: DnfPackageInfo
      description: "Info about a single package"
      properties:
        name:
          type: string
          x-go-type-skip-optional-pointer: true
          example: "nginx"
          description: Name of the package
        version:
          type: string
          x-go-type-skip-optional-pointer: true
          example: "1.18.0"
          description: Version of the package
        repo:
          type: string
          x-go-type-skip-optional-pointer: true
          example: "main"
          description: Repository from which the package was installed
      example:
        name: "nginx"
        version: "1.18.0"
        repo: "main"
    PackagesInstalled:
      title: PackagesInstalled
      type: object
      description: "List of all the installed packages"
      properties:
        packages:
          type: array
          x-go-type-skip-optional-pointer: true
          items:
            $ref: '#/components/schemas/PackageInfo'
          description: List of installed packages
      example:
        packages:
          - name: "nginx"
            version: "1.18.0"
            repo: "main"
...
    X509CertificateConfig:
      title: PkiCertificateConfig
      type: object
      description: Returns the configuration of a managed x509 certificate
      properties:
        id:
          type: string
          x-go-type-skip-optional-pointer: true
          description: Unique identifier for the certificate configuration
        role:
          type: string
          x-go-type-skip-optional-pointer: true
          description: The role associated with the certificate
        common_name:
          type: string
          x-go-type-skip-optional-pointer: true
          description: The common name (CN) for the certificate
        ttl:
          type: string
          x-go-type-skip-optional-pointer: true
          description: The time-to-live (TTL) duration for the certificate
        alt_names:
          type: array
          x-go-type-skip-optional-pointer: true
          items:
            type: string
          description: A list of alternative names (SANs) for the certificate
        ip_sans:
          type: array
          x-go-type-skip-optional-pointer: true
          items:
            type: string
          description: A list of IP SANs for the certificate
    X509CertificateData:
      title: X509Certificate
      type: object
      properties:
        issuer:
          $ref: '#/components/schemas/PkiIssuer'
        subject:
          type: string
          description: The subject of the certificate.
        serial:
          type: string
          description: The serial number of the x509 certificate.
        email_addresses:
          type: array
          x-go-type-skip-optional-pointer: true
          items:
            type: string
          description: A list of email addresses associated with the certificate.
        not_before:
          type: string
          format: date-time
          description: The start date of the certificate's validity period.
        not_after:
          type: string
          format: date-time
          description: The end date of the certificate's validity period.
        percentage:
          type: number
          x-go-type-skip-optional-pointer: true
          format: float
          description: Lifetime of the certificate.
      required:
        - issuer
        - subject
        - serial
        - not_before
        - not_after
      description: Returns the x509 certificate data
    X509ManagedCertificate:
      type: object
      title: PkiManagedCertificateConfig
      description: Represents the configuration of a managed x509 certificate
      properties:
        certificate_data:
          $ref: '#/components/schemas/X509CertificateData'
        certificate_config:
          $ref: '#/components/schemas/X509CertificateConfig'
        storage_config:
          type: array
          x-go-type-skip-optional-pointer: true
          items:
            $ref: '#/components/schemas/X509CertificateStorage'
        post_hooks:
          type: array
          x-go-type-skip-optional-pointer: true
          items:
            $ref: '#/components/schemas/PostHooks'
    X509CertificateStorage:
      type: object
      title: PkiCertificateStorage
      description: The storage configuration of a managed x509 certificate
      properties:
        cert_file:
          type: string
          x-go-type-skip-optional-pointer: true
          description: The file that the cert will be written to
        key_file:
          type: string
          x-go-type-skip-optional-pointer: true
          description: The file that the key will be written to
        ca_file:
          type: string
          x-go-type-skip-optional-pointer: true
          description: The file that the ca will be written to
    X509ManagedCertificateList:
      type: object
      title: PkiManagedCertificateList
      properties:
        data:
          type: array
          x-go-type-skip-optional-pointer: true
          items:
            $ref: '#/components/schemas/X509ManagedCertificate'
      description: Returns a list of all the configured managed PKI certificate
    PkiIssuer:
      type: object
      title: Issuer
      properties:
        serial_number:
          type: string
          description: The serial number of the issuer.
        common_name:
          type: string
          description: The common name of the issuer.
      required:
        - serial_number
        - common_name
      description: Information about the certificate's issuer.
    PostHooks:
      type: object
      title: PostHooks
      description: Hooks that are run after a cert has been issued
      properties:
        name:
          type: string
          x-go-type-skip-optional-pointer: true
          description: The name of the hook
        cmd:
          type: string
          x-go-type-skip-optional-pointer: true
          description: The command and its args that are run
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://example.com/errors/bad-request
            title: Bad Request
            status: 400
            detail: The request is invalid or missing required parameters.
        application/problem+xml:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://example.com/errors/bad-request
            title: Bad Request
            status: 400
            detail: The request is invalid or missing required parameters.
    Forbidden:
      description: Forbidden
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://example.com/errors/forbidden
            title: Forbidden
            status: 403
            detail: Access is forbidden for the provided credentials.
        application/problem+xml:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://example.com/errors/forbidden
            title: Forbidden
            status: 403
            detail: Access is forbidden for the provided credentials.
    InternalServerError:
      description: Internal Server Error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://example.com/errors/internal-server-error
            title: Internal Server Error
            status: 500
            detail: An unexpected error occurred.
        application/problem+xml:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://example.com/errors/internal-server-error
            title: Internal Server Error
            status: 500
            detail: An unexpected error occurred.
    NotFound:
      description: Not Found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://example.com/errors/not-found
            title: Not Found
            status: 404
            detail: The requested resource was not found.
        application/problem+xml:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://example.com/errors/not-found
            title: Not Found
            status: 404
            detail: The requested resource was not found.
    NotImplemented:
      description: Not Implemented
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://example.com/errors/not-implemented
            title: Not Implemented
            status: 501
            detail: This function is not enabled.
        application/problem+xml:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://example.com/errors/not-implemented
            title: Not Implemented
            status: 501
            detail: This function is not enabled.
...